# Optimized materialization of `EQ` tables