--- a/jolt-core/Cargo.toml
+++ b/jolt-core/Cargo.toml
@@ -80,3 +80,7 @@ harness = false
 [[bench]]
 name = "dense_mlpoly"
 harness = false
+
+[[bench]]
+name = "materialize"
+harness = false