--- a/jolt-core/benches/materialize.rs
+++ b/jolt-core/benches/materialize.rs
@@ -1,0 +1,20 @@
+use ark_curve25519::Fr;
+use criterion::{black_box, criterion_group, criterion_main, Criterion};
+use liblasso::jolt::subtable::eq::EqSubtable;
+use liblasso::jolt::subtable::LassoSubtable;
+
+fn materialize_subtable(c: &mut Criterion) {
+    let mut group = c.benchmark_group("DenseMLPoly Evaluation");
+
+    group.bench_function("Materialize EQ subtable", |b| {
+        b.iter(|| {
+            let table = EqSubtable::<Fr>::new();
+            black_box(table.materialize(1 << 16));
+        })
+    });
+
+    group.finish();
+}
+
+criterion_group!(benches, materialize_subtable);
+criterion_main!(benches);