--- a/jolt-core/benches/materialize.rs
+++ b/jolt-core/benches/materialize.rs
@@ -1,6 +1,7 @@
 use ark_curve25519::Fr;
 use criterion::{black_box, criterion_group, criterion_main, Criterion};
 use liblasso::jolt::subtable::eq::EqSubtable;
+use liblasso::jolt::subtable::eq_abs::EqAbsSubtable;
 use liblasso::jolt::subtable::LassoSubtable;
 
 fn materialize_subtable(c: &mut Criterion) {
@@ -9,6 +10,13 @@ fn materialize_subtable(c: &mut Criterion) {
     group.bench_function("Materialize EQ subtable", |b| {
         b.iter(|| {
             let table = EqSubtable::<Fr>::new();
+            black_box(table.materialize(1 << 16));
+        })
+    });
+
+    group.bench_function("Materialize EQAbs subtable", |b| {
+        b.iter(|| {
+            let table = EqAbsSubtable::<Fr>::new();
             black_box(table.materialize(1 << 16));
         })
     });