--- a/jolt-core/src/jolt/subtable/eq_abs.rs
+++ b/jolt-core/src/jolt/subtable/eq_abs.rs
@@ -3,7 +3,6 @@ use ark_std::log2;
 use std::marker::PhantomData;
 
 use super::LassoSubtable;
-use crate::utils::split_bits;
 
 #[derive(Default)]
 pub struct EqAbsSubtable<F: PrimeField> {