Probably in the grand scheme of things these savings are negligible, but here they are:

before:
```
DenseMLPoly Evaluation/Materialize EQ subtable
                        time:   [211.17 µs 213.41 µs 216.47 µs]

DenseMLPoly Evaluation/Materialize EQAbs subtable
                        time:   [217.81 µs 223.26 µs 229.77 µs]
```
after:
```
DenseMLPoly Evaluation/Materialize EQ subtable
                        time:   [91.029 µs 92.311 µs 93.933 µs]
                        change: [-57.152% -56.489% -55.796%] (p = 0.00 < 0.05)


DenseMLPoly Evaluation/Materialize EQAbs subtable
                        time:   [90.391 µs 91.322 µs 92.320 µs]
                        change: [-58.382% -57.365% -56.325%] (p = 0.00 < 0.05)
```