@mmagician 2024-01-03
A point against this PR would be that the `materialize` function is now slightly less readable. Arguably `EqSubtable` is still fine, while `EqAbsSubtable` got a little ugly.